Remove unneed files, reuse test code in multiple tests
