+++ test/integration/expression/lib/geometry.ts
@@ -1,29 +1,7 @@
 
 import {ICanonicalTileID, ILngLatLike} from '../../../../src';
 import {Point2D} from '../../../../src/point2d';
-
-export function mercatorXfromLng(lng: number) {
-    return (180 + lng) / 360;
-}
-
-export function mercatorYfromLat(lat: number) {
-    return (180 - (180 / Math.PI * Math.log(Math.tan(Math.PI / 4 + lat * Math.PI / 360)))) / 360;
-}
-
-function getTilePoint(canonical: ICanonicalTileID, coord: {x: number; y: number; z: number}): Point2D {
-    const tilesAtZoom = Math.pow(2, canonical.z);
-    return {
-        x: (coord.x * tilesAtZoom - canonical.x) * 8192,
-        y: (coord.y * tilesAtZoom - canonical.y) * 8192
-    };
-}
-
-function getPoint(coord: ILngLatLike, canonical: ICanonicalTileID): Point2D {
-    const p: Point2D = getTilePoint(canonical, {x: mercatorXfromLng(coord[0]), y: mercatorYfromLat(coord[1]), z: 0});
-    p.x = Math.round(p.x);
-    p.y = Math.round(p.y);
-    return p;
-}
+import {getPoint} from '../../../unit/test_utils';
 
 function convertPoint(coord: ILngLatLike, canonical: ICanonicalTileID): Point2D[] {
     return [getPoint(coord, canonical)];
